@@ -29,41 +29,45 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/core": "^5.0.0",
+    "@angular/core": "^7.0.3",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
-    "primeng": "^5.2.0",
-    "rxjs": "^5.1.0"
+    "primeng": "^7.0.0-beta.1",
+    "rxjs": "^6.3.3",
+    "tsickle": "^0.33.1",
+    "tslib": "^1.9.3",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@angular/cli": "^1.6.7",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "2.5.45",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~4.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/animations": "^7.0.3",
+    "@angular/cli": "^7.0.4",
+    "@angular/common": "^7.0.3",
+    "@angular/compiler": "^7.0.3",
+    "@angular/compiler-cli": "^7.0.3",
+    "@angular/forms": "^7.0.3",
+    "@angular/http": "^7.0.3",
+    "@angular/language-service": "^7.0.3",
+    "@angular/platform-browser": "^7.0.3",
+    "@angular/platform-browser-dynamic": "^7.0.3",
+    "@angular/router": "^7.0.3",
+    "@types/jasmine": "2.8.11",
+    "@types/node": "~10.12.3",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.3.0",
+    "jasmine-spec-reporter": "~4.2.1",
     "jquery": "^3.3.1",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "karma": "~3.1.1",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^1.7.0",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "~2.5.3",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "ng-packagr": "^4.4.0",
+    "protractor": "~5.4.1",
+    "ts-node": "~7.0.1",
+    "tslint": "~5.11.0",
+    "typescript": "^3.1.6",
     "zone.js": "^0.8.4"
   }
 }
