@@ -7,25 +7,31 @@ import {AdvancedSharedModule} from '../common/advancedshared';
 import {PrimeAdvancedTemplate} from '../common/advancedshared';
 import {AdvancedBlockableUI} from '../common/advancedblockableui';
 import {AdvancedTreeDragDropService} from '../common/advancedtreedragdropservice';
-import {Subscription} from 'rxjs/Subscription';
+import {Subscription} from 'rxjs';
 
 @Component({
     selector: 'p-atreeNode',
     template: `
         <ng-template [ngIf]="node">
-            <li *ngIf="tree.droppableNodes" class="ui-treenode-droppoint" [ngClass]="{'ui-treenode-droppoint-active ui-state-highlight':draghoverPrev}"
-            (drop)="onDropPoint($event,-1)" (dragover)="onDropPointDragOver($event)" (dragenter)="onDropPointDragEnter($event,-1)" (dragleave)="onDropPointDragLeave($event)"></li>
+            <li *ngIf="tree.droppableNodes" class="ui-treenode-droppoint"
+            [ngClass]="{'ui-treenode-droppoint-active ui-state-highlight':draghoverPrev}"
+            (drop)="onDropPoint($event,-1)" (dragover)="onDropPointDragOver($event)"
+            (dragenter)="onDropPointDragEnter($event,-1)" (dragleave)="onDropPointDragLeave($event)"></li>
             <li *ngIf="!tree.horizontal" [ngClass]="['ui-treenode',node.styleClass||'', isLeaf() ? 'ui-treenode-leaf': '']">
-                <div class="ui-treenode-content" (click)="onNodeClick($event)" (contextmenu)="onNodeRightClick($event)" (touchend)="onNodeTouchEnd()"
-                    (drop)="onDropNode($event)" (dragover)="onDropNodeDragOver($event)" (dragenter)="onDropNodeDragEnter($event)" (dragleave)="onDropNodeDragLeave($event)"
-                    [ngClass]="{'ui-treenode-selectable':tree.selectionMode && node.selectable !== false,'ui-treenode-dragover':draghoverNode, 'ui-treenode-content-selected':isSelected()}" [draggable]="tree.draggableNodes" (dragstart)="onDragStart($event)" (dragend)="onDragStop($event)">
+                <div class="ui-treenode-content" (click)="onNodeClick($event)"
+                (contextmenu)="onNodeRightClick($event)" (touchend)="onNodeTouchEnd()"
+                    (drop)="onDropNode($event)" (dragover)="onDropNodeDragOver($event)" (dragenter)="onDropNodeDragEnter($event)"
+                    (dragleave)="onDropNodeDragLeave($event)"
+                    [ngClass]="{'ui-treenode-selectable':tree.selectionMode && node.selectable !== false,
+                    'ui-treenode-dragover':draghoverNode, 'ui-treenode-content-selected':isSelected()}"
+                    [draggable]="tree.draggableNodes" (dragstart)="onDragStart($event)" (dragend)="onDragStop($event)">
                     <span class="ui-tree-toggler  fa fa-fw" [ngClass]="{'fa-caret-right':!node.expanded,'fa-caret-down':node.expanded}"
                             (click)="toggle($event)"></span
                     ><div class="ui-chkbox" *ngIf="tree.checkbox"><div class="ui-chkbox-box ui-widget ui-corner-all ui-state-default">
-                        <span class="ui-chkbox-icon ui-clickable fa" 
+                        <span class="ui-chkbox-icon ui-clickable fa"
                             [ngClass]="{'fa-check':isChecked(),'fa-minus':isPartialChecked()}"></span></div></div
                     ><span [class]="getIcon()" *ngIf="node.icon||node.expandedIcon||node.collapsedIcon"></span
-                    ><span class="ui-treenode-label ui-corner-all" 
+                    ><span class="ui-treenode-label ui-corner-all"
                         [ngClass]="{'ui-state-highlight':isSelected()}">
                             <span *ngIf="!tree.getTemplateForNode(node)" id="{{node.id}}">{{node.label}}</span>
                             <span *ngIf="tree.getTemplateForNode(node)">
@@ -33,13 +39,16 @@ import {Subscription} from 'rxjs/Subscription';
                             </span>
                     </span>
                 </div>
-                <ul class="ui-treenode-children" style="display: none;" *ngIf="node.children && node.expanded" [style.display]="node.expanded ? 'block' : 'none'">
-                    <p-atreeNode *ngFor="let childNode of node.children;let firstChild=first;let lastChild=last; let index=index" [node]="childNode" [parentNode]="node"
-                        [firstChild]="firstChild" [lastChild]="lastChild" [index]="index"></p-atreeNode>
+                <ul class="ui-treenode-children" style="display: none;" *ngIf="node.children && node.expanded"
+                [style.display]="node.expanded ? 'block' : 'none'">
+                    <p-atreeNode *ngFor="let childNode of node.children;let firstChild=first;let lastChild=last; let index=index"
+                    [node]="childNode" [parentNode]="node" [firstChild]="firstChild" [lastChild]="lastChild" [index]="index"></p-atreeNode>
                 </ul>
             </li>
-            <li *ngIf="tree.droppableNodes&&lastChild" class="ui-treenode-droppoint" [ngClass]="{'ui-treenode-droppoint-active ui-state-highlight':draghoverNext}"
-            (drop)="onDropPoint($event,1)" (dragover)="onDropPointDragOver($event)" (dragenter)="onDropPointDragEnter($event,1)" (dragleave)="onDropPointDragLeave($event)"></li>
+            <li *ngIf="tree.droppableNodes&&lastChild" class="ui-treenode-droppoint"
+            [ngClass]="{'ui-treenode-droppoint-active ui-state-highlight':draghoverNext}" (drop)="onDropPoint($event,1)"
+            (dragover)="onDropPointDragOver($event)" (dragenter)="onDropPointDragEnter($event,1)"
+            (dragleave)="onDropPointDragLeave($event)"></li>
             <table *ngIf="tree.horizontal" [class]="node.styleClass">
                 <tbody>
                     <tr>
@@ -56,24 +65,28 @@ import {Subscription} from 'rxjs/Subscription';
                             </table>
                         </td>
                         <td class="ui-treenode" [ngClass]="{'ui-treenode-collapsed':!node.expanded}">
-                            <div class="ui-treenode-content ui-state-default ui-corner-all" 
-                                [ngClass]="{'ui-treenode-selectable':tree.selectionMode,'ui-state-highlight':isSelected()}" (click)="onNodeClick($event)" (contextmenu)="onNodeRightClick($event)"
+                            <div class="ui-treenode-content ui-state-default ui-corner-all"
+                                [ngClass]="{'ui-treenode-selectable':tree.selectionMode,'ui-state-highlight':isSelected()}"
+                                (click)="onNodeClick($event)" (contextmenu)="onNodeRightClick($event)"
                                 (touchend)="onNodeTouchEnd()">
-                                <span class="ui-tree-toggler fa fa-fw" [ngClass]="{'fa-plus':!node.expanded,'fa-minus':node.expanded}" *ngIf="!isLeaf()"
+                                <span class="ui-tree-toggler fa fa-fw" [ngClass]="{'fa-plus':!node.expanded,'fa-minus':node.expanded}"
+                                *ngIf="!isLeaf()"
                                         (click)="toggle($event)"></span
                                 ><span [class]="getIcon()" *ngIf="node.icon||node.expandedIcon||node.collapsedIcon"></span
                                 ><span class="ui-treenode-label ui-corner-all">
                                         <span *ngIf="!tree.getTemplateForNode(node)" id="{{node.id}}">{{node.label}}</span>
                                         <span *ngIf="tree.getTemplateForNode(node)">
-                                        <ng-container *ngTemplateOutlet="tree.getTemplateForNode(node); context: {$implicit: node}"></ng-container>
+                                        <ng-container *ngTemplateOutlet="tree.getTemplateForNode(node);
+                                        context: {$implicit: node}"></ng-container>
                                         </span>
                                 </span>
                             </div>
                         </td>
-                        <td class="ui-treenode-children-container" *ngIf="node.children && node.expanded" [style.display]="node.expanded ? 'table-cell' : 'none'">
+                        <td class="ui-treenode-children-container" *ngIf="node.children && node.expanded"
+                        [style.display]="node.expanded ? 'table-cell' : 'none'">
                             <div class="ui-treenode-children">
-                                <p-atreeNode *ngFor="let childNode of node.children;let firstChild=first;let lastChild=last;" [node]="childNode" 
-                                        [firstChild]="firstChild" [lastChild]="lastChild"></p-atreeNode>
+                                <p-atreeNode *ngFor="let childNode of node.children;let firstChild=first;let lastChild=last;"
+                                [node]="childNode" [firstChild]="firstChild" [lastChild]="lastChild"></p-atreeNode>
                             </div>
                         </td>
                     </tr>
@@ -173,7 +186,8 @@ export class UIAdvancedTreeNode implements OnInit {
             let dropIndex = this.index;
 
             if (position < 0) {
-                dropIndex = (this.tree.dragNodeSubNodes === newNodeList) ? ((this.tree.dragNodeIndex > this.index) ? this.index : this.index - 1) : this.index;
+                dropIndex = (this.tree.dragNodeSubNodes === newNodeList) ? ((this.tree.dragNodeIndex > this.index) ?
+                this.index : this.index - 1) : this.index;
                 newNodeList.splice(dropIndex, 0, dragNode);
             }
             else {
@@ -222,7 +236,7 @@ export class UIAdvancedTreeNode implements OnInit {
 
     onDragStart(event) {
         if (this.tree.draggableNodes && this.node.draggable !== false) {
-            event.dataTransfer.setData("text", "data");
+            event.dataTransfer.setData('text', 'data');
 
             this.tree.dragDropService.startDrag({
                 tree: this,
@@ -287,7 +301,7 @@ export class UIAdvancedTreeNode implements OnInit {
     }
 
     onDropNodeDragEnter(event) {
-        if (this.tree.droppableNodes && this.node.droppable !== false 
+        if (this.tree.droppableNodes && this.node.droppable !== false
                   && this.tree.allowDrop(this.tree.dragNode, this.node, this.tree.dragNodeScope)) {
             this.draghoverNode = true;
         }
@@ -307,19 +321,21 @@ export class UIAdvancedTreeNode implements OnInit {
 @Component({
     selector: 'p-atree',
     template: `
-        <div [ngClass]="{'ui-tree ui-widget ui-widget-content ui-corner-all':true,'ui-tree-selectable':selectionMode,'ui-treenode-dragover':dragHover,'ui-tree-loading': loading}" [ngStyle]="style" [class]="styleClass" *ngIf="!horizontal"
+        <div [ngClass]="{'ui-tree ui-widget ui-widget-content ui-corner-all':true,'ui-tree-selectable':selectionMode,
+        'ui-treenode-dragover':dragHover,'ui-tree-loading': loading}" [ngStyle]="style" [class]="styleClass" *ngIf="!horizontal"
             (drop)="onDrop($event)" (dragover)="onDragOver($event)" (dragenter)="onDragEnter($event)" (dragleave)="onDragLeave($event)">
             <div class="ui-tree-loading-mask ui-widget-overlay" *ngIf="loading"></div>
             <div class="ui-tree-loading-content" *ngIf="loading">
                 <i [class]="'fa fa-spin fa-2x ' + loadingIcon"></i>
             </div>
             <ul class="ui-tree-container" *ngIf="value">
-                <p-atreeNode *ngFor="let node of value;let firstChild=first;let lastChild=last; let index=index" [node]="node" 
+                <p-atreeNode *ngFor="let node of value;let firstChild=first;let lastChild=last; let index=index" [node]="node"
                 [firstChild]="firstChild" [lastChild]="lastChild" [index]="index"></p-atreeNode>
             </ul>
             <div class="ui-tree-empty-message" *ngIf="!loading && !value">{{emptyMessage}}</div>
         </div>
-        <div [ngClass]="{'ui-tree ui-tree-horizontal ui-widget ui-widget-content ui-corner-all':true,'ui-tree-selectable':selectionMode}"  [ngStyle]="style" [class]="styleClass" *ngIf="horizontal">
+        <div [ngClass]="{'ui-tree ui-tree-horizontal ui-widget ui-widget-content ui-corner-all':true,
+        'ui-tree-selectable':selectionMode}"  [ngStyle]="style" [class]="styleClass" *ngIf="horizontal">
             <div class="ui-tree-loading ui-widget-overlay" *ngIf="loading"></div>
             <div class="ui-tree-loading-content" *ngIf="loading">
                 <i [class]="'fa fa-spin fa-2x ' + loadingIcon"></i>
@@ -879,7 +895,7 @@ export class AdvancedTree implements OnInit, AfterContentInit, OnDestroy, Advanc
 
     isPartialChecked(node: AdvancedTreeNode) {
         return this.findIndexInPartialChecked(node) !== -1;
-    }    
+    }
 
     isSingleSelectionMode() {
         return this.selectionMode && this.selectionMode === 'single';
