@@ -1,4 +1,4 @@
-import 'rxjs/add/operator/toPromise';
+
 
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
